Отдельные файлы для enum
