+++ Source/Libraries/Core/Business/VodovozBusiness/Domain/Fuel/FuelApiRequest.cs
@@ -1,6 +1,5 @@
 ﻿using QS.DomainModel.Entity;
 using System;
-using System.ComponentModel.DataAnnotations;
 using Vodovoz.Domain.Employees;
 
 namespace Vodovoz.Domain.Fuel
@@ -13,35 +12,5 @@ namespace Vodovoz.Domain.Fuel
 		public virtual FuelApiRequestFunction RequestFunction { get; set; }
 		public virtual FuelApiResponseResult ResponseResult { get; set; }
 		public virtual string ErrorResponseMessage { get; set; }
-	}
-
-	public enum FuelApiRequestFunction
-	{
-		[Display(Name = "Не указано")]
-		None,
-		[Display(Name = "Авторизация")]
-		Login,
-		[Display(Name = "Список карт")]
-		FuelCardsData,
-		[Display(Name = "Информация по карте")]
-		FuelCardDetails,
-		[Display(Name = "Список лимитов карты")]
-		FuelCardsLimitsData,
-		[Display(Name = "Удаление лимитов карты")]
-		FuelCardsLimitsDelete,
-		[Display(Name = "Создание лимита карты")]
-		FuelCardsLimitCreate,
-		[Display(Name = "Список транзакций")]
-		TransactionsData
-	}
-
-	public enum FuelApiResponseResult
-	{
-		[Display(Name = "Не указано")]
-		None,
-		[Display(Name = "Успех")]
-		Success,
-		[Display(Name = "Ошибка")]
-		Error
 	}
 }
