+++ Source/Libraries/Core/Business/VodovozBusiness/Domain/Fuel/FuelApiResponseResult.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Vodovoz.Domain.Fuel
+{
+	public enum FuelApiResponseResult
+	{
+		[Display(Name = "Не указано")]
+		None,
+		[Display(Name = "Успех")]
+		Success,
+		[Display(Name = "Ошибка")]
+		Error
+	}
+}
